+++ worlds/banjo_tooie/test/test_open_hag1.py
@@ -24,7 +24,7 @@ class TestOpenHag1WithHag1(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(55)
+        self._test_progression_jiggies(60)
 
 
 class TestOpenHag1WithBossesHag1(TestProgressionJiggies):
@@ -35,7 +35,7 @@ class TestOpenHag1WithBossesHag1(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(55)
+        self._test_progression_jiggies(60)
 
 
 class TestOpenHag1WithBosses(TestProgressionJiggies):
@@ -46,7 +46,7 @@ class TestOpenHag1WithBosses(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(55)
+        self._test_progression_jiggies(60)
 
 
 class TestClosedHag1WithHag1(TestProgressionJiggies):
@@ -57,7 +57,7 @@ class TestClosedHag1WithHag1(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(70)
+        self._test_progression_jiggies(75)
 
 
 class TestClosedHag1WithBossesHag1(TestProgressionJiggies):
@@ -68,7 +68,7 @@ class TestClosedHag1WithBossesHag1(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(55)
+        self._test_progression_jiggies(60)
 
 
 class TestClosedHag1WithBosses(TestProgressionJiggies):
@@ -79,4 +79,4 @@ class TestClosedHag1WithBosses(TestProgressionJiggies):
     }
 
     def test_progression_jiggies(self):
-        self._test_progression_jiggies(55)
+        self._test_progression_jiggies(60)